arch(#1948): migrate kommo CRM stack to src/ (slice 4)
Fourth concrete migration slice for #1948. Mirrors PR #2018 (phone_utils,
slice 1), PR #2020 (mini_app observability, slice 2), and PR #2024
(content_loader, slice 3): canonical homes move to src/, telegram_bot/
keeps thin re-export shims, mini_app's import switches to canonical paths.

Files moved (via git mv to preserve history)
--------------------------------------------
- telegram_bot/services/kommo_models.py to src/services/kommo_models.py
  Pure pydantic v2 models (LeadCreate, LeadUpdate, ContactCreate,
  TaskCreate, KommoCustomField, etc.). Module-level imports unchanged
  (only stdlib + pydantic). 261 lines.

- telegram_bot/services/kommo_tokens.py to src/services/kommo_tokens.py
  KommoTokenStore + KommoTokenStoreProtocol + Redis OAuth flow. Module-
  level imports unchanged (httpx, tenacity, stdlib). 238 lines.

- telegram_bot/services/kommo_client.py to src/services/kommo_client.py
  KommoClient + KommoOAuthAuth. Module-level imports updated:
    from telegram_bot.observability     to from src.observability
    from telegram_bot.services._retry    to from src.services._retry
    from telegram_bot.services.kommo_models  to from src.services.kommo_models
    from telegram_bot.services.kommo_tokens  to from src.services.kommo_tokens
  No telegram_bot.* references remain at module scope. 336 lines.

Files added (re-export shims)
-----------------------------
- telegram_bot/services/kommo_models.py (33 lines) re-exports every
  public model from src.services.kommo_models with object identity.
- telegram_bot/services/kommo_tokens.py (16 lines) re-exports
  KommoTokenStore, KommoTokenStoreProtocol, REDIS_KEY, REFRESH_BUFFER_SEC.
- telegram_bot/services/kommo_client.py (24 lines) re-exports
  KommoClient, KommoOAuthAuth.

Existing internal callers (telegram_bot/agents/crm_tools.py,
telegram_bot/dialogs/crm_*, telegram_bot/handlers/crm_callbacks.py,
scripts/kommo_seed.py) keep working unchanged because every public
name still resolves at the original path with the same identity.

mini_app
--------
mini_app/phone.py changes the lazy import inside get_kommo_client()
from telegram_bot.services.kommo_client to src.services.kommo_client.
Drops the last forbidden import in mini_app/phone.py per #1948.

Test patches
------------
Two tests patch httpx.AsyncClient at module scope on kommo_tokens
(test_kommo_token_store.py:56,116, test_kommo_tokens_edge_cases.py:119).
The shim does not re-export the private httpx symbol, so the patches
were repointed from telegram_bot.services.kommo_tokens.httpx to
src.services.kommo_tokens.httpx. This is the canonical mock idiom
(patch where it is defined, not where it is imported).

Drift guard
-----------
tests/contract/test_issue_1948_kommo_client_slice_contract.py asserts:

  1. All three canonical files (kommo_models, kommo_tokens,
     kommo_client) exist under src/services/.
  2. Canonical files do NOT import from telegram_bot.* at module level.
  3. Canonical modules expose the full public API (20 names total).
  4. The three telegram_bot shims re-export every public name with
     object-identity (is-equal) - same guarantee as PR #2018, PR #2024.
  5. mini_app/phone.py imports KommoClient from src.services.kommo_client
     and contains zero references to telegram_bot.services.kommo_*.

Validation
----------
  uv run pytest \
    tests/contract/test_issue_1948_kommo_client_slice_contract.py \
    tests/unit/agents/ tests/unit/dialogs/ tests/unit/services/ \
    tests/unit/handlers/ \
    -q --no-cov                                          to 2689 passed
  ruff check / ruff format --check on touched files       to clean

What does NOT change
--------------------
- No production behaviour change beyond import paths; the shims
  preserve identity (is-equal) of every public callable / class.
- KommoClient retry/refresh-on-401 semantics, the OAuth httpx.Auth
  flow, and the Redis-backed token store all keep their original
  contract verbatim.

Refs #1948 (does not close - src/api/main.py still imports
telegram_bot.graph and other heavy modules; those need separate
slices since they overlap with #1538 voice-path migration).
