+++ telegram_bot/services/kommo_models.py
@@ -1,261 +1,38 @@
-"""Pydantic v2 models for Kommo CRM API (#413).
-
-Models match Kommo API v4 payloads.
-Kommo API uses "price" for deal value; Python code uses "budget" for readability.
-Ref: https://www.kommo.com/developers/content/api/
-"""
+"""Re-export shim for Kommo pydantic models — canonical in ``src/`` (#1948 slice 4)."""
 
 from __future__ import annotations
 
-from typing import Any
-
-from pydantic import BaseModel, ConfigDict, Field
-
-
-# --- Custom field building blocks (#1655) ---
-
-
-class KommoCustomFieldValue(BaseModel):
-    """One entry inside ``KommoCustomField.values`` (Kommo API v4).
-
-    Kommo's per-field value object is ``{"value": <scalar>, "enum_code"?: str}``.
-    Modelling it explicitly removes hand-built dicts at every emitter site.
-    """
-
-    model_config = ConfigDict(populate_by_name=True)
-
-    value: Any
-    enum_code: str | None = None
-
-
-class KommoCustomField(BaseModel):
-    """Single ``custom_fields_values`` entry on a Lead/Contact payload.
-
-    Serialised shape (``model_dump(by_alias=True, exclude_none=True)``):
-
-        {"field_id": 100, "values": [{"value": "..."}, ...]}
-
-    Use :meth:`build_simple` for the common single-string-value case and
-    :meth:`dump_list` to drop ``None`` placeholders that callers emit when
-    a field id is missing from configuration.
-    """
-
-    model_config = ConfigDict(populate_by_name=True)
-
-    field_id: int
-    values: list[KommoCustomFieldValue]
-
-    @classmethod
-    def build_simple(
-        cls, *, field_id: int | None, value: Any, enum_code: str | None = None
-    ) -> KommoCustomField | None:
-        """Return a single-value field, or ``None`` if ``field_id`` is falsy.
-
-        Mirrors the existing handler guard (``if service_field_id: ...``):
-        a ``None``/``0`` field id means the CRM has not configured this
-        custom field, so emit nothing instead of pointing at field 0.
-        """
-        if not field_id:
-            return None
-        return cls(
-            field_id=field_id,
-            values=[KommoCustomFieldValue(value=value, enum_code=enum_code)],
-        )
-
-    @staticmethod
-    def dump_list(items: list[KommoCustomField | None]) -> list[dict[str, Any]]:
-        """Serialise a list of fields, skipping ``None`` placeholders.
-
-        Returns the canonical Kommo API shape so emitters can pass the
-        result straight into ``LeadCreate.custom_fields_values`` or
-        :class:`httpx.AsyncClient.post(json=...)` without further work.
-        """
-        return [
-            item.model_dump(by_alias=True, exclude_none=True) for item in items if item is not None
-        ]
-
-
-# --- Request models (Create/Update) ---
-
-
-class LeadCreate(BaseModel):
-    """POST /api/v4/leads payload."""
-
-    model_config = ConfigDict(populate_by_name=True)
-
-    name: str
-    budget: int | None = Field(None, serialization_alias="price", validation_alias="price")
-    pipeline_id: int | None = None
-    status_id: int | None = None
-    responsible_user_id: int | None = None
-    custom_fields_values: list[dict] | None = None
-
-
-class LeadUpdate(BaseModel):
-    """PATCH /api/v4/leads/{id} payload."""
-
-    model_config = ConfigDict(populate_by_name=True)
-
-    name: str | None = None
-    budget: int | None = Field(None, serialization_alias="price", validation_alias="price")
-    status_id: int | None = None
-    custom_fields_values: list[dict] | None = None
-
-
-class ContactCreate(BaseModel):
-    """POST /api/v4/contacts payload."""
-
-    first_name: str
-    last_name: str | None = None
-    phone: str | None = None
-    email: str | None = None
-    custom_fields_values: list[dict] | None = None
-
-
-class ContactUpdate(BaseModel):
-    """PATCH /api/v4/contacts/{id} payload."""
-
-    first_name: str | None = None
-    last_name: str | None = None
-    custom_fields_values: list[dict] | None = None
-
-    @staticmethod
-    def build_contact_fields(
-        phone: str | None = None,
-        email: str | None = None,
-    ) -> list[dict]:
-        """Build custom_fields_values for phone/email updates."""
-        fields = []
-        if phone is not None:
-            fields.append(
-                {"field_code": "PHONE", "values": [{"value": phone, "enum_code": "WORK"}]}
-            )
-        if email is not None:
-            fields.append(
-                {"field_code": "EMAIL", "values": [{"value": email, "enum_code": "WORK"}]}
-            )
-        return fields
-
-
-class TaskCreate(BaseModel):
-    """POST /api/v4/tasks payload."""
-
-    text: str
-    entity_id: int
-    entity_type: str = "leads"
-    complete_till: int  # Unix timestamp
-    task_type_id: int | None = None
-
-
-class TaskUpdate(BaseModel):
-    """PATCH /api/v4/tasks/{id} payload (#697)."""
-
-    text: str | None = None
-    complete_till: int | None = None  # Unix timestamp
-    responsible_user_id: int | None = None
-
-
-# --- Response models ---
-
-
-class Lead(BaseModel):
-    """Lead from Kommo API response.
-
-    Note: POST /leads returns minimal response (id only).
-    Full fields available via GET /leads/{id}.
-    Kommo API field "price" maps to "budget" in Python.
-    contacts: populated from _embedded.contacts when with=contacts requested (#731).
-    """
-
-    model_config = ConfigDict(populate_by_name=True)
-
-    id: int
-    name: str | None = None
-    budget: int | None = Field(None, validation_alias="price")
-    status_id: int | None = None
-    pipeline_id: int | None = None
-    responsible_user_id: int | None = None
-    loss_reason_id: int | None = None
-    created_at: int | None = None
-    updated_at: int | None = None
-    contacts: list[dict[str, Any]] | None = None  # from _embedded.contacts (#731)
-
-
-class Contact(BaseModel):
-    """Contact from Kommo API response."""
-
-    id: int
-    first_name: str | None = None
-    last_name: str | None = None
-    created_at: int | None = None
-
-
-class Note(BaseModel):
-    """Note from Kommo API response."""
-
-    id: int
-    text: str | None = None
-    created_at: int | None = None
-
-
-class Task(BaseModel):
-    """Task from Kommo API response."""
-
-    id: int
-    text: str | None = None
-    complete_till: int | None = None
-    entity_id: int | None = None
-    entity_type: str | None = None
-    responsible_user_id: int | None = None
-    is_completed: bool | None = None
-    # Kommo may return `result` as dict or empty list depending on task state.
-    result: dict[str, Any] | list[Any] | None = None
-    created_at: int | None = None
-    updated_at: int | None = None
-
-
-class Pipeline(BaseModel):
-    """Pipeline from Kommo API response."""
-
-    id: int
-    name: str
-    is_main: bool = False
-
-
-# --- Lead Score Sync (compatibility with existing tools/tests) ---
-
-
-class LeadScoreSyncPayload(BaseModel):
-    """Payload for syncing a lead score to Kommo custom fields."""
-
-    kommo_lead_id: int
-    score_value: int
-    score_band: str
-    score_field_id: int
-    band_field_id: int
-
-    @classmethod
-    def from_record(
-        cls,
-        rec: object,
-        *,
-        score_field_id: int,
-        band_field_id: int,
-    ) -> LeadScoreSyncPayload:
-        """Build from a LeadScoreRecord while avoiding circular imports."""
-        return cls(
-            kommo_lead_id=int(getattr(rec, "kommo_lead_id", 0) or 0),
-            score_value=int(getattr(rec, "score_value", 0)),
-            score_band=str(getattr(rec, "score_band", "")),
-            score_field_id=score_field_id,
-            band_field_id=band_field_id,
-        )
-
-    def to_kommo_payload(self) -> dict:
-        """Convert to Kommo API PATCH body for custom fields."""
-        return {
-            "custom_fields_values": [
-                {"field_id": self.score_field_id, "values": [{"value": self.score_value}]},
-                {"field_id": self.band_field_id, "values": [{"value": self.score_band}]},
-            ]
-        }
+from src.services.kommo_models import (
+    Contact,
+    ContactCreate,
+    ContactUpdate,
+    KommoCustomField,
+    KommoCustomFieldValue,
+    Lead,
+    LeadCreate,
+    LeadScoreSyncPayload,
+    LeadUpdate,
+    Note,
+    Pipeline,
+    Task,
+    TaskCreate,
+    TaskUpdate,
+)
+
+
+__all__ = [
+    "Contact",
+    "ContactCreate",
+    "ContactUpdate",
+    "KommoCustomField",
+    "KommoCustomFieldValue",
+    "Lead",
+    "LeadCreate",
+    "LeadScoreSyncPayload",
+    "LeadUpdate",
+    "Note",
+    "Pipeline",
+    "Task",
+    "TaskCreate",
+    "TaskUpdate",
+]
